+++ internal/commands/account/show_test.go
@@ -11,7 +11,7 @@ import (
 	smock "github.com/UpCloudLtd/upcloud-cli/internal/mock"
 	"github.com/UpCloudLtd/upcloud-cli/internal/output"
 
-	"github.com/UpCloudLtd/upcloud-go-api/upcloud"
+	"github.com/UpCloudLtd/upcloud-go-api/v4/upcloud"
 	"github.com/gemalto/flume"
 	"github.com/stretchr/testify/assert"
 )