+++ internal/commands/ipaddress/ip_address_test.go
@@ -3,7 +3,7 @@ package ipaddress
 import (
 	"testing"
 
-	"github.com/UpCloudLtd/upcloud-go-api/upcloud"
+	"github.com/UpCloudLtd/upcloud-go-api/v4/upcloud"
 	"github.com/stretchr/testify/assert"
 )