+++ internal/commands/networkinterface/network_interface.go
@@ -3,7 +3,7 @@ package networkinterface
 import (
 	"github.com/UpCloudLtd/upcloud-cli/internal/commands"
 	"github.com/UpCloudLtd/upcloud-cli/internal/commands/ipaddress"
-	"github.com/UpCloudLtd/upcloud-go-api/upcloud/request"
+	"github.com/UpCloudLtd/upcloud-go-api/v4/upcloud/request"
 )
 
 const maxNetworkInterfaceActions = 10