+++ CHANGELOG.md
@@ -6,6 +6,27 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- New go-api version v4.3.0
+
+### Fixed
+- Improved errors relating to argument resolver failures
+
+## [1.1.3] - 2022-02-24
+### Changed
+- Update documentation
+
+### Fixed
+- Fix storage command attached-to-server key overrides zone
+
+## [1.1.2] - 2022-01-21
+### Fixed
+- New release with no changes to fix [the Homebrew deprecation notice](https://github.com/goreleaser/goreleaser/pull/2591)
+
+## [1.1.1] - 2021-09-30
+### Changed
+- Change password creation to be disabled by default in server creation
+- Always create a password when password delivery method is chosen
 
 ## [1.1.0] - 2021-06-03
 ### Added
@@ -61,10 +82,13 @@ Initial public beta release :tada:
 ### Added
 - Current feature set added! First internal release
 
-[Unreleased]: https://github.com/UpCloudLtd/upcloud-cli/compare/v1.1.0...HEAD
+[Unreleased]: https://github.com/UpCloudLtd/upcloud-cli/compare/v1.1.3...HEAD
 [0.1.0]: https://github.com/UpCloudLtd/upcloud-cli/releases/tag/v0.1.0
 [0.1.1]: https://github.com/UpCloudLtd/upcloud-cli/releases/tag/v0.1.1
 [0.5.0]: https://github.com/UpCloudLtd/upcloud-cli/releases/tag/v0.5.0
 [0.6.0]: https://github.com/UpCloudLtd/upcloud-cli/releases/tag/v0.6.0
 [1.0.0]: https://github.com/UpCloudLtd/upcloud-cli/releases/tag/v1.0.0
 [1.1.0]: https://github.com/UpCloudLtd/upcloud-cli/releases/tag/v1.1.0
+[1.1.1]: https://github.com/UpCloudLtd/upcloud-cli/releases/tag/v1.1.1
+[1.1.2]: https://github.com/UpCloudLtd/upcloud-cli/releases/tag/v1.1.2
+[1.1.3]: https://github.com/UpCloudLtd/upcloud-cli/releases/tag/v1.1.3